refactor(date-picker): rename .scss files to .module.scss
