+++ packages/date-picker/src/DatePicker/components/DateInput/DateInput.module.scss
