+++ packages/date-picker/src/DateRangePicker/DateRangePicker.module.scss
